python-services/requirements.txt
flask==3.0.0
openai-whisper==20231117
torch>=2.0.0; sys_platform != 'darwin'
torch>=2.0.0; sys_platform == 'darwin'
numpy>=1.24.0

